% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayMean.R
\name{ArrayMean}
\alias{ArrayMean}
\title{Finds the mean of a data frame}
\usage{
ArrayMean(dataframe)
}
\arguments{
\item{dataframe}{data frame}
}
\value{
Mean of the designated data frame
}
\description{
Finds the mean of a data frame
}
\examples{
\dontrun{OutlierDetect(exampledata)}
}
