% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasedCovariance.R
\name{BiasedCovariance}
\alias{BiasedCovariance}
\title{Finds the covariance of two variables in a data frame
Must first separate columns into desired variable as shown below}
\usage{
BiasedCovariance(var1, var2)
}
\arguments{
\item{var1}{first variable}

\item{var2}{second variable}

\item{dataframe}{data frame}
}
\value{
covariance of variables
}
\description{
Finds the covariance of two variables in a data frame
Must first separate columns into desired variable as shown below
}
