% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myroots.R
\name{myroots}
\alias{myroots}
\title{Finds the correlation matrix of a two variables}
\usage{
myroots(mat)
}
\arguments{
\item{mat}{data matrix}
}
\value{
roots
}
\description{
Finds the correlation matrix of a two variables
}
