% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrot.R
\name{myrot}
\alias{myrot}
\title{Finds the mean of a data frame}
\usage{
myrot(data = mat)
}
\arguments{
\item{mat}{data matrix}
}
\value{
angle of rotation for independence, eigenvalues, eigenvectors, plot with ellipse and principle axis, orthogonality test
}
\description{
Finds the mean of a data frame
}
