% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sik.R
\name{mysik}
\alias{mysik}
\title{Displays a rotated ellipse that fits the data}
\usage{
mysik(mu, cov, n, theta)
}
\arguments{
\item{mu}{center of generated data set}

\item{cov}{Covariance Matrix}

\item{n}{number of values in generated data set}

\item{theta}{rotation in radians}
}
\value{
Generated S, S tilde, and Theta
}
\description{
Displays a rotated ellipse that fits the data
}
