% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysik2.R
\name{mysik2}
\alias{mysik2}
\title{Finds the correlation matrix of a two variables}
\usage{
mysik2(cov, theta)
}
\arguments{
\item{cov}{Covariance Matrix}

\item{theta}{rotation in radians}
}
\value{
Line 5 is the s12tilde, the graphs show the relationship between theta and s12tilde
}
\description{
Finds the correlation matrix of a two variables
}
