% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_test.R
\name{normal_mvn}
\alias{normal_mvn}
\title{Finds whether or not a data frame is consistent with multivariate normality}
\usage{
normal_mvn(x, sigma)
}
\arguments{
\item{x}{data}

\item{sigma}{number of standard deviations}
}
\value{
Barplot or result 4.29, for given number of sd
}
\description{
Finds whether or not a data frame is consistent with multivariate normality
}
