% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbvnorm.R
\name{rbvnorm}
\alias{rbvnorm}
\title{GIBBS BIVARIATE SIMULATION}
\usage{
rbvnorm(x2seed = 1, iter = 10, mu, sigma)
}
\arguments{
\item{x2seed}{seed}

\item{iter}{iterations}

\item{mu}{mu for x1 and x2}

\item{sigma}{variance/covariance matrix}
}
\value{
matrix with iter rows, x1, x2
}
\description{
GIBBS BIVARIATE SIMULATION
}
